Starting refactor of code
